LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.emHeights-low-ascent.worker-expected.txt


FAIL Testing emHeights with reduced ascent metric assert_equals: ctx.measureText('A').emHeightAscent === 20 (got 68[number], expected 20[number]) expected 20 but got 68
